refactor : 서브쿼리로 단일 쿼리에서 데이터 집계 처리

+++ src/main/java/com/example/recipeapp/domain/user/controller/dto/response/UserProfileResponseDto.java
@@ -21,7 +21,7 @@ public class UserProfileResponseDto {
     private String email;
     private int recipeCount;
     private int totalLikes;
-    private int likedRecipeCount;
+    private Long likedRecipeCount;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
